requirements.txt
reddit_bot==0.1.5
pytesseract==0.1.6
Pillow==3.0.0
